requirements.txt
Flask==2.2.3
numpy==1.23.4
opencv_python==4.6.0.66
Pillow==9.4.0
tensorflow==2.12.0rc0
torch==2.0.0
torchvision==0.15.1
